+++ app/server.js
@@ -31,7 +31,7 @@ app.use(webpackHotMiddleware(compiler, {
 app.use(express.static('./public'));
 
 app.get('/hello', function (req, res) {
-    res.send('Hello, world!');
+    res.send('Index, world!');
 });
 
 app.post("/register", routes.insert);
