+++ public/entry.jsx
@@ -1,9 +1,9 @@
-import Index from './index.jsx';
+import App from './app.jsx';
 import OrderPage from './component/order.jsx';
 import Rent from './component/rent.jsx';
 import GoodsDetails from './component/rent-details.jsx';
 import SignIn from './component/sign-in.jsx'
-import Hello from './hello.jsx';
+import Index from './index.jsx';
 import IndexRent from './component/index-rent.jsx';
 import Register from './component/register.jsx';
 import React from 'react';
@@ -14,14 +14,15 @@ require('jquery');
 require("bootstrap-webpack");
 
 const router = <Router history={hashHistory}>
-    <Route path="/" component={Index}>
-        <IndexRoute component={Hello}/>
+    <Route path="/" component={App}>
+        <IndexRoute component={Index}/>
         <Route path='/login' component={SignIn}/>
         <Route path='/register' component={Register}/>
         <Route path='/indexRent' component={IndexRent}/>
         <Route path='/rent' component={Rent}/>
         <Route path='/goodsDetails' component={GoodsDetails}/>
         <Route path='/orderPage' component={OrderPage}/>
+        <Route path='/personalPage' />
     </Route>
 </Router>;
