+++ public/index.jsx
@@ -1,28 +1,26 @@
 import React from 'react';
-import LoginAndRegisterNavs from './component/login-and-register-nav.jsx';
-import TabsNavs from './component/navs.jsx';
-
-
-export default class Index extends React.Component {
+import BrifeIntroduciton from './component/brief-introduction.jsx';
+import PicturesWall from './component/pictures-wall.jsx';
+import Buttom from './component/buttom.jsx';
+export default class Hello extends React.Component {
     render() {
         return (
-            <div className="container-fluid ">
-                <div className="container-fluid">
+            <div className="container-fluid">
+                <div>
                     <div>
-                        <div className="conta">
-                            <TabsNavs/>
-                        </div>
-                        <div className="pull-right loginNavs">
-                            <LoginAndRegisterNavs/>
-                        </div>
+                        <PicturesWall/>
+                    </div>
+
+                    <div>
+                        <BrifeIntroduciton/>
                     </div>
                 </div>
-
                 <div>
                     {this.props.children}
                 </div>
-
-
+                <div className="buttom">
+                    <Buttom/>
+                </div>
             </div>
         );
     }
